refactor: make MainActivity for navbar
