+++ app/src/main/java/com/team22/soundary/feature/profile/ProfileEditedFragment.kt
@@ -1,0 +1,106 @@
+package com.team22.soundary.feature.main.profile
+
+import android.Manifest
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.os.Build
+import android.os.Bundle
+import android.provider.MediaStore
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.Fragment
+import com.team22.soundary.databinding.FragmentMypageEditBinding
+
+class ProfileEditedFragment : Fragment() {
+    private var _binding: FragmentMypageEditBinding? = null
+    private val binding get() = _binding!!
+    private val PERMISSION_REQUEST_CODE = 100
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentMypageEditBinding.inflate(inflater,container,false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding.profileImageview.setOnClickListener {
+            if (checkAndRequestPermissions()) {
+                openGallery()
+            }
+        }
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
+
+    private fun checkAndRequestPermissions(): Boolean {
+        val permissionsNeeded = mutableListOf<String>()
+
+        // 갤러리 접근
+        val readPermission = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            ContextCompat.checkSelfPermission(requireContext(), Manifest.permission.READ_MEDIA_IMAGES)
+        } else {
+            ContextCompat.checkSelfPermission(requireContext(), Manifest.permission.READ_EXTERNAL_STORAGE)
+        }
+
+        if (readPermission != PackageManager.PERMISSION_GRANTED) {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+                permissionsNeeded.add(Manifest.permission.READ_MEDIA_IMAGES)
+            } else {
+                permissionsNeeded.add(Manifest.permission.READ_EXTERNAL_STORAGE)
+            }
+        }
+
+        // 권한 요청 필요 시 처리
+        return if (permissionsNeeded.isNotEmpty()) {
+            ActivityCompat.requestPermissions(
+                requireActivity(),
+                permissionsNeeded.toTypedArray(),
+                PERMISSION_REQUEST_CODE
+            )
+            false
+        } else {
+            true
+        }
+    }
+
+    // 갤러리 열기
+    private fun openGallery() {
+        val intent = Intent(Intent.ACTION_PICK, MediaStore.Images.Media.EXTERNAL_CONTENT_URI)
+        startActivityForResult(intent, PERMISSION_REQUEST_CODE)
+    }
+
+
+    //TODO : Fragment 변경에 맞게 아래 메서드 수정 필요해 보입니다.
+
+//    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+//        super.onActivityResult(requestCode, resultCode, data)
+//        if (requestCode == PERMISSION_REQUEST_CODE && resultCode == RESULT_OK && data != null) {
+//            val selectedImage: Uri? = data.data
+//            binding.profileImageView.setImageURI(selectedImage)  // 선택된 이미지 표시
+//        }
+//    }
+//
+//    // 권한 요청 결과 처리
+//    override fun onRequestPermissionsResult(
+//        requestCode: Int,
+//        permissions: Array<String>,
+//        grantResults: IntArray
+//    ) {
+//        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
+//        if (requestCode == PERMISSION_REQUEST_CODE) {
+//            if (grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+//                openGallery()
+//            }
+//        }
+//    }
+}
