--- app/src/main/java/com/team22/soundary/feature/profile/ProflieEditedActivity.kt
@@ -1,100 +1,0 @@
-package com.team22.soundary.feature.main.profile
-
-import android.Manifest
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.net.Uri
-import android.os.Build
-import android.os.Bundle
-import android.provider.MediaStore
-import android.widget.Button
-import android.widget.EditText
-import android.widget.ImageView
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
-import com.team22.soundary.R
-
-class ProfileEditActivity : AppCompatActivity() {
-
-    private val PERMISSION_REQUEST_CODE = 100
-    private lateinit var profileImageView: ImageView
-    private lateinit var profileIntroEdit: EditText
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_mypage_edit)
-
-        profileImageView = findViewById(R.id.profile_imageview)
-        profileIntroEdit = findViewById(R.id.profile_intro_edit)
-
-        profileImageView.setOnClickListener {
-            if (checkAndRequestPermissions()) {
-                openGallery()
-            }
-        }
-
-        val buttonHiphop: Button = findViewById(R.id.button_hiphop)
-        buttonHiphop.setOnClickListener {
-        }
-    }
-
-    private fun checkAndRequestPermissions(): Boolean {
-        val permissionsNeeded = mutableListOf<String>()
-
-        // 갤러리 접근
-        val readPermission = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            ContextCompat.checkSelfPermission(this, Manifest.permission.READ_MEDIA_IMAGES)
-        } else {
-            ContextCompat.checkSelfPermission(this, Manifest.permission.READ_EXTERNAL_STORAGE)
-        }
-
-        if (readPermission != PackageManager.PERMISSION_GRANTED) {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-                permissionsNeeded.add(Manifest.permission.READ_MEDIA_IMAGES)
-            } else {
-                permissionsNeeded.add(Manifest.permission.READ_EXTERNAL_STORAGE)
-            }
-        }
-
-        // 권한 요청 필요 시 처리
-        return if (permissionsNeeded.isNotEmpty()) {
-            ActivityCompat.requestPermissions(
-                this,
-                permissionsNeeded.toTypedArray(),
-                PERMISSION_REQUEST_CODE
-            )
-            false
-        } else {
-            true
-        }
-    }
-
-    // 갤러리 열기
-    private fun openGallery() {
-        val intent = Intent(Intent.ACTION_PICK, MediaStore.Images.Media.EXTERNAL_CONTENT_URI)
-        startActivityForResult(intent, PERMISSION_REQUEST_CODE)
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == PERMISSION_REQUEST_CODE && resultCode == RESULT_OK && data != null) {
-            val selectedImage: Uri? = data.data
-            profileImageView.setImageURI(selectedImage)  // 선택된 이미지 표시
-        }
-    }
-
-    // 권한 요청 결과 처리
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if (requestCode == PERMISSION_REQUEST_CODE) {
-            if (grantResults.isNotEmpty() && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                openGallery()
-            }
-        }
-    }
-}
