+++ app/src/main/res/drawable/nav_friend_ic.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M0,20V17.225C0,16.658 0.142,16.133 0.425,15.65C0.708,15.167 1.1,14.8 1.6,14.55C1.833,14.433 2.058,14.325 2.275,14.225C2.508,14.125 2.75,14.033 3,13.95V20H0ZM4,13C3.167,13 2.458,12.708 1.875,12.125C1.292,11.542 1,10.833 1,10C1,9.167 1.292,8.458 1.875,7.875C2.458,7.292 3.167,7 4,7C4.833,7 5.542,7.292 6.125,7.875C6.708,8.458 7,9.167 7,10C7,10.833 6.708,11.542 6.125,12.125C5.542,12.708 4.833,13 4,13ZM4,11C4.283,11 4.517,10.908 4.7,10.725C4.9,10.525 5,10.283 5,10C5,9.717 4.9,9.483 4.7,9.3C4.517,9.1 4.283,9 4,9C3.717,9 3.475,9.1 3.275,9.3C3.092,9.483 3,9.717 3,10C3,10.283 3.092,10.525 3.275,10.725C3.475,10.908 3.717,11 4,11ZM4,20V17.2C4,16.633 4.142,16.117 4.425,15.65C4.725,15.167 5.117,14.8 5.6,14.55C6.633,14.033 7.683,13.65 8.75,13.4C9.817,13.133 10.9,13 12,13C13.1,13 14.183,13.133 15.25,13.4C16.317,13.65 17.367,14.033 18.4,14.55C18.883,14.8 19.267,15.167 19.55,15.65C19.85,16.117 20,16.633 20,17.2V20H4ZM6,18H18V17.2C18,17.017 17.95,16.85 17.85,16.7C17.767,16.55 17.65,16.433 17.5,16.35C16.6,15.9 15.692,15.567 14.775,15.35C13.858,15.117 12.933,15 12,15C11.067,15 10.142,15.117 9.225,15.35C8.308,15.567 7.4,15.9 6.5,16.35C6.35,16.433 6.225,16.55 6.125,16.7C6.042,16.85 6,17.017 6,17.2V18ZM12,12C10.9,12 9.958,11.608 9.175,10.825C8.392,10.042 8,9.1 8,8C8,6.9 8.392,5.958 9.175,5.175C9.958,4.392 10.9,4 12,4C13.1,4 14.042,4.392 14.825,5.175C15.608,5.958 16,6.9 16,8C16,9.1 15.608,10.042 14.825,10.825C14.042,11.608 13.1,12 12,12ZM12,10C12.55,10 13.017,9.808 13.4,9.425C13.8,9.025 14,8.55 14,8C14,7.45 13.8,6.983 13.4,6.6C13.017,6.2 12.55,6 12,6C11.45,6 10.975,6.2 10.575,6.6C10.192,6.983 10,7.45 10,8C10,8.55 10.192,9.025 10.575,9.425C10.975,9.808 11.45,10 12,10ZM20,13C19.167,13 18.458,12.708 17.875,12.125C17.292,11.542 17,10.833 17,10C17,9.167 17.292,8.458 17.875,7.875C18.458,7.292 19.167,7 20,7C20.833,7 21.542,7.292 22.125,7.875C22.708,8.458 23,9.167 23,10C23,10.833 22.708,11.542 22.125,12.125C21.542,12.708 20.833,13 20,13ZM20,11C20.283,11 20.517,10.908 20.7,10.725C20.9,10.525 21,10.283 21,10C21,9.717 20.9,9.483 20.7,9.3C20.517,9.1 20.283,9 20,9C19.717,9 19.475,9.1 19.275,9.3C19.092,9.483 19,9.717 19,10C19,10.283 19.092,10.525 19.275,10.725C19.475,10.908 19.717,11 20,11ZM21,20V13.95C21.25,14.033 21.483,14.125 21.7,14.225C21.933,14.325 22.167,14.433 22.4,14.55C22.9,14.8 23.292,15.167 23.575,15.65C23.858,16.133 24,16.658 24,17.225V20H21Z"
+      android:fillColor="#1D1B20"/>
+</vector>
