+++ app/src/main/res/layout/fragment_friend_profile.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.search.FriendProfileActivity">
+    tools:context=".feature.search.FriendProfileFragment">
 
     <ImageView
         android:id="@+id/back_icon"
