+++ app/src/main/res/layout/fragment_friend_search.xml
