+++ app/src/main/res/menu/all_menus.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:id="@+id/action_main"
+        android:icon="@drawable/nav_home_ic"
+        android:title="홈"/>
+    <item android:id="@+id/action_share"
+        android:icon="@drawable/main_share_icon"
+        android:title="공유"/>
+    <item android:id="@+id/action_friend"
+        android:icon="@drawable/nav_friend_ic"
+        android:title="친구"/>
+    <item android:id="@+id/action_my"
+        android:icon="@drawable/nav_my_ic"
+        android:title="내 정보"/>
+</menu>
