+++ app/src/main/res/values/dimens.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<resources>
+<resources xmlns:tools="http://schemas.android.com/tools">
     <!-- all -->
+    <dimen name="design_bottom_navigation_height" tools:override="true">62dp</dimen>
     <!-- margin -->
     <dimen name="margin_half">8dp</dimen>
     <dimen name="margin_normal">16dp</dimen>
