+++ app/src/main/res/values/strings.xml
@@ -89,5 +89,7 @@
     <string name="genre_Jpop">JPOP</string>
 
     <string name="add_friend">친구 추가</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
