+++ app/src/main/res/xml/motion_scene_main.xml
@@ -1,0 +1,101 @@
+<?xml version="1.0" encoding="utf-8"?>
+<MotionScene xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:motion="http://schemas.android.com/apk/res-auto">
+
+    <Transition
+        android:id="@+id/next_transition"
+        motion:constraintSetEnd="@id/next"
+        motion:constraintSetStart="@id/start"
+        motion:duration="600">
+    </Transition>
+
+    <ConstraintSet android:id="@+id/start">
+        <Constraint android:id="@id/current_imageView"
+            android:layout_width="@dimen/music_cover_size"
+            android:layout_height="@dimen/music_cover_size"
+            android:background="@drawable/ic_launcher_background"
+            motion:layout_constraintBottom_toBottomOf="parent"
+            motion:layout_constraintEnd_toEndOf="parent"
+            motion:layout_constraintStart_toStartOf="parent"
+            motion:layout_constraintTop_toTopOf="parent"
+            motion:layout_constraintVertical_bias="0.2" />
+        <Constraint android:id="@+id/next_imageView"
+            android:layout_width="@dimen/music_cover_size"
+            android:layout_height="@dimen/music_cover_size"
+            android:layout_marginTop="110dp"
+            android:background="@drawable/main_music_cover"
+            android:rotation="20"
+            motion:layout_constraintBottom_toBottomOf="@id/current_imageView"
+            motion:layout_constraintStart_toEndOf="parent"
+            motion:layout_constraintTop_toTopOf="@id/current_imageView"  />
+        <Constraint android:id="@+id/prev_imageView"
+            android:layout_width="@dimen/music_cover_size"
+            android:layout_height="@dimen/music_cover_size"
+            android:layout_marginTop="110dp"
+            android:background="@drawable/main_music_cover"
+            android:rotation="-20"
+            motion:layout_constraintBottom_toBottomOf="@id/current_imageView"
+            motion:layout_constraintEnd_toStartOf="parent"
+            motion:layout_constraintTop_toTopOf="@id/current_imageView" />
+    </ConstraintSet>
+
+    <ConstraintSet android:id="@+id/next">
+        <Constraint
+            android:id="@id/current_imageView"
+            android:layout_width="@dimen/music_cover_size"
+            android:layout_height="@dimen/music_cover_size"
+            android:layout_marginTop="110dp"
+            android:rotation="-20"
+            motion:layout_constraintBottom_toTopOf="@+id/message_textView"
+            motion:layout_constraintEnd_toStartOf="parent" />
+
+        <Constraint
+            android:id="@id/next_imageView"
+            android:layout_width="@dimen/music_cover_size"
+            android:layout_height="@dimen/music_cover_size"
+            android:rotation="0"
+            motion:layout_constraintBottom_toBottomOf="parent"
+            motion:layout_constraintEnd_toEndOf="parent"
+            motion:layout_constraintStart_toStartOf="parent"
+            motion:layout_constraintTop_toTopOf="parent"
+            motion:layout_constraintVertical_bias="0.2" />
+
+        <Constraint
+            android:id="@id/prev_imageView"
+            motion:layout_constraintStart_toStartOf="parent" />
+    </ConstraintSet>
+
+    <Transition
+        android:id="@+id/prev_transition"
+        motion:constraintSetEnd="@id/prev"
+        motion:constraintSetStart="@id/start"
+        motion:duration="600">
+    </Transition>
+
+    <ConstraintSet android:id="@+id/prev">
+        <Constraint
+            android:id="@id/current_imageView"
+            android:layout_width="@dimen/music_cover_size"
+            android:layout_height="@dimen/music_cover_size"
+            android:layout_marginTop="110dp"
+            android:rotation="20"
+            motion:layout_constraintStart_toEndOf="parent"
+            motion:layout_constraintBottom_toTopOf="@+id/message_textView" />
+
+        <Constraint
+            android:id="@id/next_imageView"
+            motion:layout_constraintEnd_toEndOf="parent" />
+
+        <Constraint
+            android:id="@id/prev_imageView"
+            android:layout_width="@dimen/music_cover_size"
+            android:layout_height="@dimen/music_cover_size"
+            android:rotation="0"
+            motion:layout_constraintBottom_toBottomOf="parent"
+            motion:layout_constraintEnd_toEndOf="parent"
+            motion:layout_constraintStart_toStartOf="parent"
+            motion:layout_constraintTop_toTopOf="parent"
+            motion:layout_constraintVertical_bias="0.2"/>
+    </ConstraintSet>
+
+</MotionScene>
